+++ test/explicit/type/date_time_posix_test.rb
@@ -15,6 +15,30 @@ class Explicit::Type::DateTimePosixTest < ActiveSupport::TestCase
     )
   end
 
+  test "min" do
+    assert_ok(
+      "2024-12-11 11:37:43".to_datetime,
+      validate(1733917063, [:date_time_posix, min: -> { "2024-12-11 11:37:43".to_datetime }])
+    )
+
+    assert_error(
+      "must not be a moment before 2024-12-11T11:37:43+00:00",
+      validate(1733917062, [:date_time_posix, min: -> { "2024-12-11 11:37:43".to_datetime }])
+    )
+  end
+
+  test "max" do
+    assert_ok(
+      "2024-12-11 11:37:43".to_datetime,
+      validate(1733917063, [:date_time_posix, max: -> { "2024-12-11 11:37:43".to_datetime }])
+    )
+
+    assert_error(
+      "must not be a moment after 2024-12-11T11:37:43+00:00",
+      validate(1733917064, [:date_time_posix, max: -> { "2024-12-11 11:37:43".to_datetime }])
+    )
+  end
+
   test "error" do
     assert_error "must be a valid posix timestamp", validate("foo", :date_time_posix)
     assert_error "must be a valid posix timestamp", validate(nil, :date_time_posix)